+++ src/data/data/products.ts
@@ -8,6 +8,7 @@ import {
   PRODUCT_KIT,
   CUSTOMER_PARAMETERS,
   IN_STOCK,
+  CHEVRONS,
 } from 'src/data/constants';
 import { Cordura, Xpac } from './fabrics/fabrics';
 
@@ -15,7 +16,7 @@ const bikeWallet = () => ({
   id: 'bike-wallet',
   name: 'BikeWallet',
   volume: null,
-  size: `13/8`,
+  size: `13 x 8`,
   weight: null,
   material: null,
   description: {
@@ -43,7 +44,7 @@ const organizer = () => ({
   id: 'organizer',
   name: 'Organizer',
   volume: null,
-  size: `18/12/5`,
+  size: `18 x 12 x 5`,
   weight: null,
   material: null,
   description: {
@@ -68,7 +69,7 @@ const backpackForCity = () => ({
   id: 'backpack-for-city',
   name: 'BackpackForCity',
   volume: null,
-  size: '40/25/12',
+  size: '40 x 25 x 12',
   sizeOfTheClosedBag: true,
   weight: null,
   material: null,
@@ -94,7 +95,7 @@ const hamster = () => ({
   id: 'hamster',
   name: 'Hamster',
   volume: '1',
-  size: `22/9/10`,
+  size: `22 x 9 x 10`,
   weight: [
     {
       fabric: 'Cordura',
@@ -128,7 +129,7 @@ const barrelBagWithoutValve = () => ({
   id: 'barrel-bag-without-valve',
   name: 'BarrelBagWithoutValve',
   volume: '3',
-  size: `14/25`,
+  size: `14 x 25`,
   weight: [
     {
       fabric: 'Cordura',
@@ -162,7 +163,7 @@ const seatBag3L = () => ({
   id: 'seat-bag-3l',
   name: 'SeatBag3l',
   volume: '3',
-  size: `11/12/30`,
+  size: `11 x 12 x 30`,
   weight: [
     {
       fabric: 'Cordura',
@@ -236,7 +237,7 @@ const triangularBicycleFirstAidKitFrame = () => ({
   id: 'triangular-bicycle-first-aid-kit-frame',
   name: 'TriangularBicycleFirstAidKitFrame',
   volume: '0.5',
-  size: `18/15/6`,
+  size: `18 x 15 x 6`,
   weight: [
     {
       fabric: 'Cordura',
@@ -270,7 +271,7 @@ const smallBagTubeOnBar = () => ({
   id: 'small-bag-tube-on-bar',
   name: 'SmallBagTubeOnBar',
   volume: '4',
-  size: `24/21/14`,
+  size: `24 x 21 x 14`,
   weight: [
     {
       fabric: 'Cordura',
@@ -304,7 +305,7 @@ const bananaOnBar = () => ({
   id: 'banana-on-bar',
   name: 'BananaOnBar',
   volume: '4',
-  size: `17/33/14`,
+  size: `17 x 33 x 14`,
   weight: [
     {
       fabric: 'Cordura',
@@ -338,7 +339,7 @@ const seatBag5L = () => ({
   id: 'seat-bag-5l',
   name: 'SeatBag5l',
   volume: '5',
-  size: `16/17/38`,
+  size: `16 x 17 x 38`,
   weight: [
     {
       fabric: 'Cordura',
@@ -406,7 +407,7 @@ const barrelBagWithValve = () => ({
   id: 'barrel-bag-with-valve',
   name: 'BarrelBagWithValve',
   volume: '3',
-  size: `14/25`,
+  size: `14 x 25`,
   weight: [
     {
       fabric: 'Cordura',
@@ -476,7 +477,7 @@ const ovalBarrelbag = () => ({
   id: 'oval-barrelbag',
   name: 'OvalBarrelbag',
   volume: '3.5',
-  size: `19/14/25`,
+  size: `19 x 14 x 25`,
   weight: [
     {
       fabric: 'Cordura',
@@ -510,7 +511,7 @@ const smallToolSeatbag = () => ({
   id: 'small-tool-seatbag',
   name: 'SmallToolSeatbag',
   volume: '0.8',
-  size: `21/5-7/11`,
+  size: `21 x 5-7 x 11`,
   weight: null,
   material: null,
   description: {
@@ -535,7 +536,7 @@ const barrelBagMini = () => ({
   id: 'barrel-bag-mini',
   name: 'BarrelBagMini',
   volume: '0.75',
-  size: '19.5/8',
+  size: '19.5 x 8',
   weight: [
     {
       fabric: 'Cordura',
@@ -569,7 +570,7 @@ const tacticalStembag = () => ({
   id: 'tactical-stembag',
   name: 'TacticalStembag',
   volume: '1.0',
-  size: '22/9/10',
+  size: '22 x 9 x 10',
   weight: [
     {
       fabric: 'Cordura',
@@ -628,7 +629,7 @@ const barrelbagWithVelcro = () => ({
   id: 'barrelbag-with-velcro',
   name: 'BarrelbagWithVelcro',
   volume: '3',
-  size: '14/25',
+  size: '14 x 25',
   weight: [
     {
       fabric: 'Cordura',
@@ -662,7 +663,7 @@ const phonePouch = () => ({
   id: 'phone-pouch',
   name: 'PhonePouch',
   volume: null,
-  size: '17/8',
+  size: '17 x 8',
   weight: null,
   material: null,
   description: {
@@ -687,7 +688,7 @@ const universalBagForCity = () => ({
   id: 'universal-bag-for-city',
   name: 'UniversalBagForCity',
   volume: '1',
-  size: '12/9/20',
+  size: '12 x 9 x 20',
   weight: null,
   material: null,
   description: {
@@ -724,7 +725,7 @@ const handlebarMount = () => ({
   price: { UAH: 800, EUR: 25 },
   fabrics: null,
   embedVideo: { id: 'ylSSGONW8Hc', title: 'HandlebarMountVideoTitle' },
-  category: BICYCLE_EQUIPMENT,
+  category: BAG_ACCESSORIES,
   subcategory: SUBCATEGORIES_BICYCLE_EQUIPMENT.Mounts,
   productKit: null,
   createdAt: '2023-12-22',
@@ -734,7 +735,7 @@ const handlebarMount = () => ({
 //   id: 'hamster-xpac-without-pockets',
 //   name: 'Hamster',
 //   volume: '1',
-//   size: `22/9/10`,
+//   size: `22 x 9 x 10`,
 //   weight: [
 //     {
 //       fabric: 'Xpac',
@@ -762,7 +763,7 @@ const hamsterCorduraNavyBlueFoxes = () => ({
   id: 'hamster-cordura-navyblue-foxes',
   name: 'Hamster',
   volume: '1',
-  size: `22/9/10`,
+  size: `22 x 9 x 10`,
   weight: [
     {
       fabric: 'Cordura',
@@ -790,7 +791,7 @@ const hamsterCorduraBlackEmbroidery = () => ({
   id: 'hamster-cordura-black-embroidery',
   name: 'Hamster',
   volume: '0.7',
-  size: `16/9/10`,
+  size: `16 x 9 x 10`,
   weight: [
     {
       fabric: 'Cordura',
@@ -818,7 +819,7 @@ const halfFrameBagXpacOrange = () => ({
   id: 'half-frame-bag-xpac-orange',
   name: 'HalfFrameBag',
   volume: '~ 1',
-  size: `10/30/6`,
+  size: `10 x 30 x 6`,
   weight: [
     {
       fabric: 'Xpac',
@@ -846,7 +847,7 @@ const halfFrameBagXpacYellowBlue = () => ({
   id: 'half-frame-bag-xpac-yellow-blue',
   name: 'HalfFrameBag',
   volume: '~ 1',
-  size: `10/30/6`,
+  size: `10 x 30 x 6`,
   weight: [
     {
       fabric: 'Xpac',
@@ -874,7 +875,7 @@ const halfFrameBagXpacYellowBlue = () => ({
 //   id: 'barrel-bag-with-embroidery-blue',
 //   name: 'BarrelBagWithEmbroidery',
 //   volume: '3',
-//   size: `14/25`,
+//   size: `14 x 25`,
 //   weight: [
 //     {
 //       fabric: 'Cordura',
@@ -902,7 +903,7 @@ const barrelBagWithEmbroideryYellow = () => ({
   id: 'barrel-bag-with-embroidery-yellow',
   name: 'BarrelBagWithEmbroidery',
   volume: '3',
-  size: `14/25`,
+  size: `14 x 25`,
   weight: [
     {
       fabric: 'Cordura',
@@ -924,6 +925,214 @@ const barrelBagWithEmbroideryYellow = () => ({
   subcategory: SUBCATEGORIES_BICYCLE_EQUIPMENT.HandlebarBags,
   productKit: PRODUCT_KIT,
   createdAt: '2024-07-11',
+});
+
+const donutChevron = () => ({
+  id: 'donut',
+  name: 'ChevronDonut',
+  volume: null,
+  size: '7 x 7.5',
+  weight: null,
+  material: null,
+  description: {
+    short: 'ChevronDonutDescription',
+    main: 'ChevronUsage',
+    conclusion: null,
+  },
+  mounts: 'ChevronMountType',
+  gallery: 3,
+  price: {
+    UAH: 160,
+    EUR: 4,
+  },
+  fabrics: null,
+  embedVideo: null,
+  category: CHEVRONS,
+  subcategory: null,
+  productKit: null,
+  createdAt: '2025-03-20',
+});
+
+const ghostChevron = () => ({
+  id: 'ghost',
+  name: 'ChevronGhost',
+  volume: null,
+  size: '6.5 x 8',
+  weight: null,
+  material: null,
+  description: {
+    short: 'ChevronGhostDescription',
+    main: 'ChevronUsage',
+    conclusion: null,
+  },
+  mounts: 'ChevronMountType',
+  gallery: 5,
+  price: {
+    UAH: 90,
+    EUR: 2,
+  },
+  fabrics: null,
+  embedVideo: null,
+  category: CHEVRONS,
+  subcategory: null,
+  productKit: null,
+  createdAt: '2025-03-20',
+});
+
+const heartChevron = () => ({
+  id: 'heart',
+  name: 'ChevronHeart',
+  volume: null,
+  size: '6 x 7',
+  weight: null,
+  material: null,
+  description: {
+    short: 'ChevronHeartDescription',
+    main: 'ChevronUsage',
+    conclusion: null,
+  },
+  mounts: 'ChevronMountType',
+  gallery: 3,
+  price: {
+    UAH: 120,
+    EUR: 3,
+  },
+  fabrics: null,
+  embedVideo: null,
+  category: CHEVRONS,
+  subcategory: null,
+  productKit: null,
+  createdAt: '2025-03-20',
+});
+
+const coffeeChevron = () => ({
+  id: 'coffee',
+  name: 'ChevronCoffee',
+  volume: null,
+  size: '6.5 x 3',
+  weight: null,
+  material: null,
+  description: {
+    short: 'ChevronCoffeeDescription',
+    main: 'ChevronUsage',
+    conclusion: null,
+  },
+  mounts: 'ChevronMountType',
+  gallery: 4,
+  price: {
+    UAH: 90,
+    EUR: 3,
+  },
+  fabrics: null,
+  embedVideo: null,
+  category: CHEVRONS,
+  subcategory: null,
+  productKit: null,
+  createdAt: '2025-03-20',
+});
+
+const letsgoChevron = () => ({
+  id: 'letsgo',
+  name: 'ChevronLetsGo',
+  volume: null,
+  size: '6.5 x 3',
+  weight: null,
+  material: null,
+  description: {
+    short: 'ChevronLetsGoDescription',
+    main: 'ChevronUsage',
+    conclusion: null,
+  },
+  mounts: 'ChevronMountType',
+  gallery: 4,
+  price: {
+    UAH: 90,
+    EUR: 3,
+  },
+  fabrics: null,
+  embedVideo: null,
+  category: CHEVRONS,
+  subcategory: null,
+  productKit: null,
+  createdAt: '2025-03-20',
+});
+
+const sliceChevron = () => ({
+  id: 'slice',
+  name: 'ChevronSlice',
+  volume: null,
+  size: '4.5 x 4.5',
+  weight: null,
+  material: null,
+  description: {
+    short: 'ChevronSliceDescription',
+    main: 'ChevronUsage',
+    conclusion: null,
+  },
+  mounts: 'ChevronMountType',
+  gallery: 3,
+  price: {
+    UAH: 70,
+    EUR: 2,
+  },
+  fabrics: null,
+  embedVideo: null,
+  category: CHEVRONS,
+  subcategory: null,
+  productKit: null,
+  createdAt: '2025-03-20',
+});
+
+const stravaChevron = () => ({
+  id: 'strava',
+  name: 'ChevronStrava',
+  volume: null,
+  size: '6.5 x 3',
+  weight: null,
+  material: null,
+  description: {
+    short: 'ChevronStravaDescription',
+    main: 'ChevronUsage',
+    conclusion: null,
+  },
+  mounts: 'ChevronMountType',
+  gallery: 4,
+  price: {
+    UAH: 90,
+    EUR: 3,
+  },
+  fabrics: null,
+  embedVideo: null,
+  category: CHEVRONS,
+  subcategory: null,
+  productKit: null,
+  createdAt: '2025-03-20',
+});
+
+const veloCatChevron = () => ({
+  id: 'velo-cat',
+  name: 'ChevronVelocats',
+  volume: null,
+  size: '7 x 8',
+  weight: null,
+  material: null,
+  description: {
+    short: 'ChevronVelocatsDescription',
+    main: 'ChevronUsage',
+    conclusion: null,
+  },
+  mounts: 'ChevronMountType',
+  gallery: 3,
+  price: {
+    UAH: 130,
+    EUR: 4,
+  },
+  fabrics: null,
+  embedVideo: null,
+  category: CHEVRONS,
+  subcategory: null,
+  productKit: null,
+  createdAt: '2025-03-20',
 });
 
 const Products: () => Product[] = () => [
@@ -957,6 +1166,14 @@ const Products: () => Product[] = () => [
   halfFrameBagXpacYellowBlue(),
   // barrelBagWithEmbroideryBlue(),
   barrelBagWithEmbroideryYellow(),
+  donutChevron(),
+  ghostChevron(),
+  heartChevron(),
+  coffeeChevron(),
+  letsgoChevron(),
+  sliceChevron(),
+  stravaChevron(),
+  veloCatChevron(),
 ];
 
 export {
@@ -990,6 +1207,14 @@ export {
   halfFrameBagXpacYellowBlue,
   // barrelBagWithEmbroideryBlue,
   barrelBagWithEmbroideryYellow,
+  donutChevron,
+  ghostChevron,
+  heartChevron,
+  coffeeChevron,
+  letsgoChevron,
+  sliceChevron,
+  stravaChevron,
+  veloCatChevron,
 };
 
 export default Products;